Porting to gradle cont.  Working out proguard and 65k method issues
