+++ library/src/androidTest/AndroidManifest.xml
