+++ library/src/androidTest/java/org/jboss/aerogear/android/authorization/test/helper/Data.java
