+++ library/src/androidTest/java/org/jboss/aerogear/android/authorization/test/util/StubActivity.java
