+++ library/src/androidTest/res/values/strings.xml
